+++ changelog.md
@@ -4,6 +4,7 @@
 - [FIX][IOS] Unwrapp AsyncImage URL and replace placeholder 
 - [FIX][IOS] Add function preferenceTapped to catalogViewToolbarTemplate
 - [FIX][AND] Fix catalog Empty state background color not overridable
+- [FEA][AND] add entrie line templates in basket preview 
 
 # 3.2.0
 - [FIX][AND] Add missing templates in myMeal and item selector